+++ src/app/mod.rs
@@ -16,7 +16,6 @@
     clippy::similar_names,
     clippy::struct_excessive_bools,
     clippy::struct_field_names,
-    clippy::too_many_lines,
     clippy::trivially_copy_pass_by_ref,
     clippy::unused_self
 )]